app/requirements.txt
anyio==3.6.2
boto3==1.26.122
botocore==1.29.122
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
einops==0.6.1
encodec==0.1.1
fastapi==0.95.1
filelock==3.12.0
fsspec==2023.4.0
funcy==2.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
s3transfer==0.6.0
scipy==1.10.1
soundfile==0.12.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
suno-bark @ git+https://github.com/suno-ai/bark.git@2c12023eb22868a633b76357b69d657b374736d9
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.2
